+++ CHANGELOG.md
@@ -130,3 +130,4 @@
 - [2026-04-04 19:45] feat(i18n): add bi-directional event bus handler for i18n
 - [2026-04-04 20:28] feat(test): add bi-directional event bus handler for test
 - [2026-04-04 20:30] fix(state): resolve race condition during state initialization
+- [2026-04-04 20:37] test(network): add unit tests for network business logic